Rework isCandidate function to address potential race condition, that might
have been observed with an EnvoyFilter not being deleted as as the
transflect deployment annotation has been removed.

Change isCandidate name to shouldProcess, invoked from within process
function under deployment lock - ie only one replicaset operation at a
time.

It might worth while considering a bigger refactor observing Deployments,
rather than replicasets.

Add missing invocation of Deployment informer for cleanup.

Move MutexMap to transflect package to keep the main package lean.

This merges the following commits:
* operator: Use same informer factory
* operator: Move MutexMap to transflect package
* Rework isCandidate function

     Makefile                            |   2 +-
     cmd/transflect-operator/mutexmap.go |  25 ------
     cmd/transflect-operator/operator.go | 120 +++++++++++++++-------------
     pkg/transflect/mutexmap.go          |  29 +++++++
     4 files changed, 94 insertions(+), 82 deletions(-)

Pull-Request: https://github.com/cashapp/transflect/pull/5

+++ cmd/transflect-operator/operator.go
@@ -9,6 +9,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/cashapp/transflect/pkg/transflect"
 	"github.com/jpillora/backoff"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
@@ -41,9 +42,10 @@ type operator struct {
 	//
 	// 	    activeState[deploymentKey] = { revision, grpcPort }
 	activeState sync.Map
+
 	// serialise all operations on a deployment,
 	// so that no concurrent operations on a single deployment are possible
-	deploymentLocker mutexMap
+	deploymentLocker transflect.MutexMap
 
 	useIngress bool
 	plaintext  bool
@@ -62,11 +64,13 @@ func newOperator(cfg *config) (*operator, error) {
 		return nil, err
 	}
 
-	rsInformer := informers.NewSharedInformerFactory(k8s, time.Second*30).Apps().V1().ReplicaSets()
+	informerFactory := informers.NewSharedInformerFactory(k8s, time.Second*30)
+
+	rsInformer := informerFactory.Apps().V1().ReplicaSets()
 	if err := rsInformer.Informer().SetWatchErrorHandler(watchErrorHandler); err != nil {
 		return nil, errors.Wrap(err, "cannot add custom error handler to replicaset informer")
 	}
-	deployInformer := informers.NewSharedInformerFactory(k8s, time.Second*30).Apps().V1().Deployments()
+	deployInformer := informerFactory.Apps().V1().Deployments()
 	if err := deployInformer.Informer().SetWatchErrorHandler(watchErrorHandler); err != nil {
 		return nil, errors.Wrap(err, "cannot add custom error handler to deployment informer")
 	}
@@ -108,6 +112,7 @@ func (o *operator) start() error {
 	log.Debug().Msg("Starting: informer event handlers registered")
 	o.runWorkers(42)
 	o.rsInformer.Informer().Run(o.stopper)
+	o.deployInformer.Informer().Run(o.stopper)
 
 	// When run finishes e.g. because of signal calling o.stop():
 	log.Debug().Msg("Shutting down queue")
@@ -171,13 +176,26 @@ func (o *operator) enqueue(v interface{}) {
 		log.Error().Interface("object", v).Msg("Cannot convert object to Replicaset for enqueueing")
 		return
 	}
-	if key, ok := o.candidateKey(rs); ok {
-		o.queue.Add(key)
+	if !shouldEnqueue(rs) {
+		return
 	}
+	key, _ := cache.MetaNamespaceKeyFunc(rs)
+	o.queue.Add(key)
 }
 
 func (o *operator) enqueueWithOld(_, v interface{}) {
 	o.enqueue(v)
+}
+
+func shouldEnqueue(rs *appsv1.ReplicaSet) bool {
+	if rs.Status.ReadyReplicas == 0 {
+		return false
+	}
+	if _, ok := getDeploymentKey(rs); !ok {
+		log.Debug().Str("replica", rs.Name).Msg("Replicaset does not have deployment owner")
+		return false
+	}
+	return true
 }
 
 func (o *operator) cleanupDeployment(v interface{}) {
@@ -192,50 +210,7 @@ func (o *operator) cleanupDeployment(v interface{}) {
 		return
 	}
 	o.activeState.Delete(key)
-	o.deploymentLocker.delete(key)
-}
-
-func (o *operator) candidateKey(rs *appsv1.ReplicaSet) (string, bool) {
-	if !o.isCandidate(rs) {
-		return "", false
-	}
-	rsKey, _ := cache.MetaNamespaceKeyFunc(rs)
-	return rsKey, true
-}
-
-func (o *operator) isCandidate(rs *appsv1.ReplicaSet) bool {
-	if rs.Status.ReadyReplicas == 0 {
-		return false
-	}
-	deployKey, ok := getDeploymentKey(rs)
-	if !ok {
-		log.Error().Str("replica", rs.Name).Msg("Cannot get deployment Key for candidate evaluation")
-		return false
-	}
-	port := grpcPort(rs)
-	v, existing := o.activeState.Load(deployKey)
-	if existing {
-		active, _ := v.(activeEntry)
-		// Ignore candidate if we have an existing filter but the candidate is
-		// old or has not changed the port.
-		revision := deployRevision(rs)
-		if revision == 0 {
-			log.Error().Str("replica", rs.Name).Msg("Cannot get revision annotation for candidate evaluation")
-			return false
-		}
-		if revision < active.revision {
-			return false
-		}
-		if revision == active.revision && port == active.grpcPort {
-			return false
-		}
-	}
-	if !existing && port == 0 {
-		// No existing record, so nothing to do if not annotated
-		return false
-	}
-	// Upsert or delete EnvoyFilter for Replicaset
-	return true
+	o.deploymentLocker.Remove(key)
 }
 
 func (o *operator) runWorkers(cnt int) {
@@ -273,9 +248,6 @@ func (o *operator) next() bool {
 		log.Error().Err(err).Msg("Cannot get next queued Replicaset")
 		return true
 	}
-	if !o.isCandidate(rs) {
-		return true
-	}
 	if err := o.processFilter(rs); err != nil {
 		log.Error().Err(err).Str("replica", rs.Name).Msg("Cannot process EnvoyFilter for Replicaset")
 	}
@@ -284,9 +256,11 @@ func (o *operator) next() bool {
 
 func (o *operator) processFilter(rs *appsv1.ReplicaSet) error {
 	deployKey, _ := getDeploymentKey(rs)
-	revision := deployRevision(rs)
-	o.deploymentLocker.lock(deployKey)
-	defer o.deploymentLocker.unlock(deployKey)
+	o.deploymentLocker.Lock(deployKey)
+	defer o.deploymentLocker.Unlock(deployKey)
+	if !o.shouldProcess(rs) {
+		return nil
+	}
 	port := grpcPort(rs)
 	if port == 0 {
 		if err := o.deleteFilter(context.Background(), rs); err != nil {
@@ -296,16 +270,50 @@ func (o *operator) processFilter(rs *appsv1.ReplicaSet) error {
 			log.Warn().Err(err).Str("replica", rs.Name).Msg("Cannot delete EnvoyFilter because it cannot be found")
 		}
 		o.activeState.Delete(deployKey)
-		o.deploymentLocker.delete(deployKey)
+		o.deploymentLocker.Remove(deployKey)
 		return nil
 	}
 
 	if err := o.upsertFilter(context.Background(), rs); err != nil {
 		return err
 	}
+	revision := deployRevision(rs)
 	active := activeEntry{grpcPort: port, revision: revision}
 	o.activeState.Store(deployKey, active)
 	return nil
+}
+
+func (o *operator) shouldProcess(rs *appsv1.ReplicaSet) bool {
+	if rs.Status.ReadyReplicas == 0 {
+		return false
+	}
+	deployKey, _ := getDeploymentKey(rs)
+	port := grpcPort(rs)
+	v, existing := o.activeState.Load(deployKey)
+	if existing {
+		active, _ := v.(activeEntry)
+		// Ignore candidate if we have an existing filter but the candidate is
+		// old or has not changed the port.
+		revision := deployRevision(rs)
+		if revision == 0 {
+			log.Error().Str("replica", rs.Name).Msg("Cannot get revision annotation for candidate evaluation")
+			return false
+		}
+		if revision < active.revision {
+			return false
+		}
+		// Potential race condition if not performed under deployment
+		// lock and port is changed several times in a short period of time.
+		if revision == active.revision && port == active.grpcPort {
+			return false
+		}
+	}
+	if !existing && port == 0 {
+		// No existing record, so nothing to do if not annotated
+		return false
+	}
+	// Upsert or delete EnvoyFilter for Replicaset
+	return true
 }
 
 func (o *operator) getReplicaset(key string) (*appsv1.ReplicaSet, error) {
